Move getColor into colorFunc
